Prepare benchmark further

+++ pisinger_reader.h
@@ -1,7 +1,7 @@
 #ifndef PISINGER_READER_H
 #define PISINGER_READER_H
 
-void pisinger_reader(int *n, int *c, int *z, int **p, int **w, int **x,
+void pisinger_reader(int *n, int *c, long *z, int **p, int **w, int **x,
                      char *problem_file, int problem_no);
 
 #endif
